home.html에 테스트데이터생성링크 추가, mypage 추가

+++ src/main/java/com/cbg/exam/repository/MemberRepository.java
@@ -30,7 +30,7 @@ public class MemberRepository{
         return save(member);
     }
 
-    public Member findById(int id){
+    public Member findById(Long id){
         return em.find(Member.class, id);
     }
